+++ src/emulator/functionsEmulatorRuntime.ts
@@ -1018,30 +1018,12 @@ async function main(): Promise<void> {
   });
   app.all(`/*`, async (req: express.Request, res: express.Response) => {
     try {
-      new EmulatorLog(
-        "INFO",
-        "runtime-status",
-        `Beginning execution of "${FUNCTION_TARGET_NAME}"`
-      ).log();
-
       const trigger = FUNCTION_TARGET_NAME.split(".").reduce((mod, functionTargetPart) => {
         return mod?.[functionTargetPart];
       }, functionModule) as CloudFunction<unknown>;
       if (!trigger) {
         throw new Error(`Failed to find function ${FUNCTION_TARGET_NAME} in the loaded module`);
       }
-
-      const startHrTime = process.hrtime();
-      res.on("finish", () => {
-        const elapsedHrTime = process.hrtime(startHrTime);
-        new EmulatorLog(
-          "INFO",
-          "runtime-status",
-          `Finished "${FUNCTION_TARGET_NAME}" in ${
-            elapsedHrTime[0] * 1000 + elapsedHrTime[1] / 1000000
-          }ms`
-        ).log();
-      });
 
       switch (FUNCTION_SIGNATURE) {
         case "event":
@@ -1063,25 +1045,9 @@ async function main(): Promise<void> {
       res.status(500).send(err.message);
     }
   });
-  const server = app.listen(process.env.PORT, () => {
+  app.listen(process.env.PORT, () => {
     logDebug(`Listening to port: ${process.env.PORT}`);
   });
-  if (!FUNCTION_DEBUG_MODE) {
-    let timeout = process.env.FUNCTIONS_EMULATOR_TIMEOUT_SECONDS || "60";
-    if (timeout.endsWith("s")) {
-      timeout = timeout.slice(0, -1);
-    }
-    const timeoutMs = parseInt(timeout, 10) * 1000;
-    server.setTimeout(timeoutMs, () => {
-      new EmulatorLog(
-        "FATAL",
-        "runtime-error",
-        `Your function timed out after ~${timeout}s. To configure this timeout, see
-      https://firebase.google.com/docs/functions/manage-functions#set_timeout_and_memory_allocation.`
-      ).log();
-      return flushAndExit(1);
-    });
-  }
 
   // Event emitters do not work well with async functions, so we
   // construct our own promise chain to make sure each message is